+++ lib/rubanok/plane.rb
@@ -30,10 +30,10 @@ module Rubanok
   #
   # You can access the input data via `raw` method.
   class Plane
-    class << self
-      include DSL::Mapping
-      include DSL::Matching
+    include DSL::Matching
+    include DSL::Mapping
 
+    class << self
       def call(input, params)
         new(input).call(params)
       end